Allow SQL Statements to be sorted
Another important part of this project is that Statements need to be
executed in the right order when constructing a sqlite database. I had
some big plans for this; I wanted to put Table names and inter-table
dependencies in the Statement type, but thought better of it, and
decided to keep Statements leaner, and store those things at a higher
level.