Give input and output structure
This project is essentially a converter between formats, so my idea is
to have an intermediate data format, and then a series of input and
output components. The intermediate data format is implemented by
Database, Table and Statement, so now I'm formalising what an input and an
output looks like with these files.