Order INSERTS by their first element
I noticed some weird diff noise when using this tool because the order of
rows would change when creating plain-text sql output. I think ordering
by the first value of the insert statement will basically work for my
purposes, but I might like to extend this later to read the primary key
from the CREATE statement, instead of assuming it's the first column.