Implement inputDBFile
This function will allow the user to read directly from a sqlite
database without needing to dump from the sqlite binary.

I'm not super happy with this implementation, I wrestled with it for
ages, and I feel it's pretty rough. I'm committing it now though,
because it's working for my purposes and I'd rather start with something
basic and iterate than polish endlessly.